+++ TinkoffASDKUI/TinkoffASDKUITests/Infrastructure/Mocks/MaskedTextFieldDelegateListenerMock.swift
@@ -1,0 +1,36 @@
+//
+//  MaskedTextFieldDelegateListenerMock.swift
+//  TinkoffASDKUI-Unit-Tests
+//
+//  Created by Ivan Glushko on 01.06.2023.
+//
+
+@testable import TinkoffASDKUI
+
+final class MaskedTextFieldDelegateListenerMock: NSObject, MaskedTextFieldDelegateListener {
+
+    // MARK: - textField
+
+    typealias TextFieldArguments = (textField: UITextField, complete: Bool, value: String)
+
+    var textFieldCallsCount = 0
+    var textFieldReceivedArguments: TextFieldArguments?
+    var textFieldReceivedInvocations: [TextFieldArguments] = []
+
+    @objc
+    func textField(_ textField: UITextField, didFillMask complete: Bool, extractValue value: String) {
+        textFieldCallsCount += 1
+        let arguments = (textField, complete, value)
+        textFieldReceivedArguments = arguments
+        textFieldReceivedInvocations.append(arguments)
+    }
+}
+
+extension MaskedTextFieldDelegateListenerMock {
+
+    func fullReset() {
+        textFieldCallsCount = .zero
+        textFieldReceivedArguments = nil
+        textFieldReceivedInvocations = []
+    }
+}