+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [EACQAPW-4924] BankResolver and PaymentSystemResolver Tests
 * [EACQAPW-4929] TinkoffPaySheetPresenter Tests
 * [EACQAPW-4932] TinkoffPay Tests
+* [EACQAPW-4193] Add Card Tests
 
 ### Fixed